+++ crates/cargo-contract/src/cmd/extrinsics/mod.rs
@@ -46,7 +46,13 @@ use std::{
     path::PathBuf,
 };
 
-use crate::DEFAULT_KEY_COL_WIDTH;
+use crate::{
+    cmd::{
+        Balance,
+        Client,
+    },
+    DEFAULT_KEY_COL_WIDTH,
+};
 
 use contract_build::{
     name_value_println,
@@ -91,10 +97,7 @@ pub use remove::RemoveCommand;
 pub use subxt::PolkadotConfig as DefaultConfig;
 pub use upload::UploadCommand;
 
-type Balance = u128;
-type CodeHash = <DefaultConfig as Config>::Hash;
 type PairSigner = tx::PairSigner<DefaultConfig, sr25519::Pair>;
-type Client = OnlineClient<DefaultConfig>;
 
 /// Arguments required for creating and sending an extrinsic to a substrate node.
 #[derive(Clone, Debug, clap::Args)]